Properly memoise values from `Thunk`s
The side effect of getting this to work is to make the Thunk mutable,
which means that the caller must mark a Thunk as `mut` in order to
properly force it.

The alternative is to return a new `Thunk` when forcing that has the new
stored values, and providing a `get` or similarly named function in
order to retrieve the value in question e.g.

```
pub fn force(&self) -> Thunk<T, F> { <snip> }
pub fn value(&self) -> &T { <snip> }
```

or equivalent; I only specify &T as I suppose it makes sense to return
an immutable reference to the value rather than copying it.