Finish Range port
This does some tomfoolery for mimicking GeneralizedNewtypeDeriving that
can probably be dropped one day. There is also some savagery in a few of
the implementations that can be cleaned up with a future audit.

I'll consider this finished when I've got a 1:1 port of all the
functions from `range.rs` in here.